Remove separate tests directory (#26) Move tests to new files in the same directory as the implementation files they target. Put test fixture classes in test files, rather than helper file.
